move arrowtrap creation method into its own file
