+++ lib/spawning/create.lua
@@ -281,40 +281,6 @@ function module.create_spikeball(x, y, l)
 	spikeballlib.create_spikeball_trap(x, y, l)
 end
 
-function module.create_arrowtrap(x, y, l)
-	-- local entity_here = get_grid_entity_at(x, y, l)
-	-- if entity_here ~= -1 then
-    --     -- get_entity(entity_here):destroy()
-	-- 	kill_entity(entity_here)
-	-- end
-	removelib.remove_floor_and_embedded_at(x, y, l)
-    local uid = spawn_grid_entity(ENT_TYPE.FLOOR_ARROW_TRAP, x, y, l)
-    local left = validlib.is_solid_grid_entity(x-1, y, l)
-    local right = validlib.is_solid_grid_entity(x+1, y, l)
-	local flip = false
-	if not left and not right then
-		--math.randomseed(read_prng()[5])
-		if prng:random() < 0.5 then
-			flip = true
-		end
-	elseif not left then
-		flip = true
-	end
-	if flip == true then
-		flip_entity(uid)
-	end
-	if test_flag(state.level_flags, 18) == true then
-		spawn_entity_over(ENT_TYPE.FX_SMALLFLAME, uid, 0, 0.35)
-	end
-
-	if state.theme == THEME.CITY_OF_GOLD then
-		local texture_def = get_texture_definition(TEXTURE.DATA_TEXTURES_FLOORMISC_0)
-		texture_def.texture_path = "res/floormisc_gold_trap.png"
-		get_entity(uid):set_texture(define_texture(texture_def))
-	end
-
-end
-
 function module.create_giantspider(x, y, l)
     spawn_entity(ENT_TYPE.MONS_GIANTSPIDER, x+.5, y, l, 0, 0)
     module.GIANTSPIDER_SPAWNED = true
